refac: move ranks list to consts and improve rank controller
