+++ backend/src/consts/Rank.ts
@@ -1,0 +1,1 @@
+export const allRanks = ['Bronze', 'Silver', 'Gold', 'Platinum', 'Diamond', 'Master', 'Grandmaster'];
